Factor out database models
