+++ server/database/index.js
@@ -1,26 +1,4 @@
-const mongoose = require('mongoose');
-const connection = require('./config');
-const Schema = mongoose.Schema;
-
-const ProductSchema = new Schema({
-  id: Number,
-  name: String,
-  price: Number,
-  category: String,
-  img1_url: String
-});
-
-const CartItemSchema = new Schema({
-  id: Number,
-  name: String,
-  price: Number,
-  category: String,
-  img1_url: String,
-  quantity: Number,
-});
-
-const CartItems = connection.model('cart_item', CartItemSchema);
-const Products = connection.model('product', ProductSchema);
+const { CartItems, Products } = require('./models');
 
 // cart methods
 const cart = {
